feat/#39: 히스토리 저장 로직 수정, 사이즈 추천 로직 훅으로 분리
