+++ src/components/common/Header/Header.tsx
@@ -18,7 +18,7 @@ function Header(props: HeaderProps) {
   const { removeLocalStorageItem } = useRemoveLocalStorage();
 
   const reset = () => {
-    removeLocalStorageItem(['currentView', 'currentTab', 'recommend-size', 'productImage', 'history']);
+    removeLocalStorageItem('currentView', 'currentTab', 'recommend-size', 'productImage', 'history');
     window.close();
   };
