+++ src/components/common/Layout.tsx
@@ -18,12 +18,12 @@ function Layout(props: PropsWithChildren<LayoutProps>) {
   const currentView = useRecoilValue(currentViewState);
   const history = useRecoilValue(historyState);
 
-  if (history) {
-    localStorage.setItem('history', history);
-  }
-  localStorage.setItem('currentView', currentView);
-
   useEffect(() => {
+    if (history) {
+      localStorage.setItem('history', history);
+    }
+    localStorage.setItem('currentView', currentView);
+
     if (currentView !== 'compare') {
       localStorage.removeItem('currentTab');
     }
