+++ src/hooks/queries/useRemoveLocalStorage.ts
@@ -1,6 +1,8 @@
 export const useRemoveLocalStorage = () => {
   const removeLocalStorageItem = (...args: unknown[]) => {
-    args.forEach((key) => localStorage.removeItem(`${key}`));
+    args.forEach((key) => {
+      localStorage.removeItem(`${key}`);
+    });
   };
 
   return { removeLocalStorageItem };
