+++ src/pages/Popup/save-product/index.tsx
@@ -6,11 +6,11 @@ import Main from '../../../components/common/Main';
 import SizeRecommendButton from '../../../components/save-product/SizeRecommendButton';
 import { DOMAIN } from '../../../contants/domain';
 import { LINK, MESSAGE } from '../../../contants/text';
-import { productState, topOrBottomState } from '../../../states/atom';
+import { productState, sizeRecommendState } from '../../../states/atom';
 import theme from '../../../styles/theme';
 
 function SaveProduct() {
-  const topOrBottom = useRecoilValue(topOrBottomState);
+  const sizeRecommend = useRecoilValue(sizeRecommendState);
 
   const getLink = <Styled.Link onClick={() => window.open(DOMAIN.HOME)}>{LINK.SAVE}</Styled.Link>;
   const { image } = useRecoilValue(productState);
@@ -24,7 +24,7 @@ function SaveProduct() {
         link={getLink}
         noPadding
       />
-      {!topOrBottom && <SizeRecommendButton />}
+      {!sizeRecommend && <SizeRecommendButton />}
     </Layout>
   );
 }
